move decorators in new .py file
Make State class as Singleton, with context_manager's dunder methods
